+++ src/posix/enumerate.rs
@@ -8,18 +8,20 @@ cfg_if! {
 
 cfg_if! {
     if #[cfg(any(target_os = "ios", target_os = "macos"))] {
-        use core_foundation_sys::base::*;
-        use core_foundation_sys::dictionary::*;
-        use core_foundation_sys::number::*;
-        use core_foundation_sys::string::*;
+        use core_foundation::base::CFType;
+        use core_foundation::base::TCFType;
+        use core_foundation::dictionary::CFDictionary;
+        use core_foundation::dictionary::CFMutableDictionary;
+        use core_foundation::number::CFNumber;
+        use core_foundation::string::CFString;
+        use core_foundation_sys::base::{kCFAllocatorDefault, CFRetain};
         use io_kit_sys::*;
         use io_kit_sys::keys::*;
         use io_kit_sys::serial::keys::*;
         use io_kit_sys::types::*;
         use io_kit_sys::usb::lib::*;
         use nix::libc::{c_char, c_void};
-        use std::convert::TryInto;
-        use std::ffi::{CStr, CString};
+        use std::ffi::CStr;
         use std::mem::MaybeUninit;
     }
 }
@@ -31,11 +33,7 @@ cfg_if! {
     target_os = "macos"
 ))]
 use crate::SerialPortType;
-#[cfg(any(
-    target_os = "ios",
-    all(target_os = "linux", not(target_env = "musl"), feature = "libudev"),
-    target_os = "macos"
-))]
+#[cfg(any(target_os = "ios", target_os = "linux", target_os = "macos"))]
 use crate::UsbPortInfo;
 #[cfg(any(
     target_os = "android",
@@ -140,9 +138,6 @@ fn port_type(d: &libudev::Device) -> Result<SerialPortType> {
             let usb_properties = vec![
                 d.property_value("ID_USB_VENDOR_ID"),
                 d.property_value("ID_USB_MODEL_ID"),
-                d.property_value("ID_USB_VENDOR"),
-                d.property_value("ID_USB_MODEL"),
-                d.property_value("ID_USB_SERIAL_SHORT"),
             ]
             .into_iter()
             .collect::<Option<Vec<_>>>();
@@ -177,6 +172,7 @@ fn port_type(d: &libudev::Device) -> Result<SerialPortType> {
                 Ok(SerialPortType::PciPort)
             }
         }
+        None if is_rfcomm(d) => Ok(SerialPortType::BluetoothPort),
         None => find_usb_interface_from_parents(d.parent())
             .and_then(get_modalias_from_device)
             .as_deref()
@@ -299,91 +295,54 @@ fn get_parent_device_by_type(
 #[cfg(any(target_os = "ios", target_os = "macos"))]
 #[allow(non_upper_case_globals)]
 /// Returns a specific property of the given device as an integer.
-fn get_int_property(
-    device_type: io_registry_entry_t,
-    property: &str,
-    cf_number_type: CFNumberType,
-) -> Option<u32> {
-    unsafe {
-        let prop_str = CString::new(property).unwrap();
-        let key = CFStringCreateWithCString(
+fn get_int_property(device_type: io_registry_entry_t, property: &str) -> Result<u32> {
+    let cf_property = CFString::new(property);
+
+    let cf_type_ref = unsafe {
+        IORegistryEntryCreateCFProperty(
+            device_type,
+            cf_property.as_concrete_TypeRef(),
             kCFAllocatorDefault,
-            prop_str.as_ptr(),
-            kCFStringEncodingUTF8,
-        );
-        let _key_guard = scopeguard::guard((), |_| {
-            CFRelease(key as *const c_void);
-        });
-
-        let container = IORegistryEntryCreateCFProperty(device_type, key, kCFAllocatorDefault, 0);
-        if container.is_null() {
-            return None;
-        }
-        let _container_guard = scopeguard::guard((), |_| {
-            CFRelease(container);
-        });
-
-        match cf_number_type {
-            kCFNumberSInt8Type => {
-                let mut num: u8 = 0;
-                let num_ptr: *mut c_void = &mut num as *mut _ as *mut c_void;
-                CFNumberGetValue(container as CFNumberRef, cf_number_type, num_ptr);
-                Some(num as u32)
-            }
-            kCFNumberSInt16Type => {
-                let mut num: u16 = 0;
-                let num_ptr: *mut c_void = &mut num as *mut _ as *mut c_void;
-                CFNumberGetValue(container as CFNumberRef, cf_number_type, num_ptr);
-                Some(num as u32)
-            }
-            kCFNumberSInt32Type => {
-                let mut num: u32 = 0;
-                let num_ptr: *mut c_void = &mut num as *mut _ as *mut c_void;
-                CFNumberGetValue(container as CFNumberRef, cf_number_type, num_ptr);
-                Some(num)
-            }
-            _ => None,
-        }
+            0,
+        )
+    };
+    if cf_type_ref.is_null() {
+        return Err(Error::new(ErrorKind::Unknown, "Failed to get property"));
     }
+
+    let cf_type = unsafe { CFType::wrap_under_create_rule(cf_type_ref) };
+    cf_type
+        .downcast::<CFNumber>()
+        .and_then(|n| n.to_i64())
+        .map(|n| n as u32)
+        .ok_or(Error::new(
+            ErrorKind::Unknown,
+            "Failed to get numerical value",
+        ))
 }
 
 #[cfg(any(target_os = "ios", target_os = "macos"))]
 /// Returns a specific property of the given device as a string.
-fn get_string_property(device_type: io_registry_entry_t, property: &str) -> Option<String> {
-    unsafe {
-        let prop_str = CString::new(property).unwrap();
-        let key = CFStringCreateWithCString(
+fn get_string_property(device_type: io_registry_entry_t, property: &str) -> Result<String> {
+    let cf_property = CFString::new(property);
+
+    let cf_type_ref = unsafe {
+        IORegistryEntryCreateCFProperty(
+            device_type,
+            cf_property.as_concrete_TypeRef(),
             kCFAllocatorDefault,
-            prop_str.as_ptr(),
-            kCFStringEncodingUTF8,
-        );
-        let _key_guard = scopeguard::guard((), |_| {
-            CFRelease(key as *const c_void);
-        });
-
-        let container = IORegistryEntryCreateCFProperty(device_type, key, kCFAllocatorDefault, 0);
-        if container.is_null() {
-            return None;
-        }
-        let _container_guard = scopeguard::guard((), |_| {
-            CFRelease(container);
-        });
-
-        let mut buf = Vec::with_capacity(256);
-        let result = CFStringGetCString(
-            container as CFStringRef,
-            buf.as_mut_ptr(),
-            buf.capacity().try_into().unwrap(),
-            kCFStringEncodingUTF8,
-        );
-        let opt_str = if result != 0 {
-            CStr::from_ptr(buf.as_ptr()).to_str().ok().map(String::from)
-        } else {
-            None
-        };
-
-        opt_str
+            0,
+        )
+    };
+    if cf_type_ref.is_null() {
+        return Err(Error::new(ErrorKind::Unknown, "Failed to get property"));
     }
+
+    let cf_type = unsafe { CFType::wrap_under_create_rule(cf_type_ref) };
+    cf_type
+        .downcast::<CFString>()
+        .map(|s| s.to_string())
+        .ok_or(Error::new(ErrorKind::Unknown, "Failed to get string value"))
 }
 
 #[cfg(any(target_os = "ios", target_os = "macos"))]
@@ -398,13 +357,11 @@ fn port_type(service: io_object_t) -> SerialPortType {
         .or_else(|| get_parent_device_by_type(service, legacy_usb_device_class_name));
     if let Some(usb_device) = maybe_usb_device {
         SerialPortType::UsbPort(UsbPortInfo {
-            vid: get_int_property(usb_device, "idVendor", kCFNumberSInt16Type).unwrap_or_default()
-                as u16,
-            pid: get_int_property(usb_device, "idProduct", kCFNumberSInt16Type).unwrap_or_default()
-                as u16,
-            serial_number: get_string_property(usb_device, "USB Serial Number"),
-            manufacturer: get_string_property(usb_device, "USB Vendor Name"),
-            product: get_string_property(usb_device, "USB Product Name"),
+            vid: get_int_property(usb_device, "idVendor").unwrap_or_default() as u16,
+            pid: get_int_property(usb_device, "idProduct").unwrap_or_default() as u16,
+            serial_number: get_string_property(usb_device, "USB Serial Number").ok(),
+            manufacturer: get_string_property(usb_device, "USB Vendor Name").ok(),
+            product: get_string_property(usb_device, "USB Product Name").ok(),
             // Apple developer documentation indicates `bInterfaceNumber` is the supported key for
             // looking up the composite usb interface id. `idVendor` and `idProduct` are included in the same tables, so
             // we will lookup the interface number using the same method. See:
@@ -412,8 +369,9 @@ fn port_type(service: io_object_t) -> SerialPortType {
             // https://developer.apple.com/documentation/bundleresources/entitlements/com_apple_developer_driverkit_transport_usb
             // https://developer.apple.com/library/archive/documentation/DeviceDrivers/Conceptual/USBBook/USBOverview/USBOverview.html#//apple_ref/doc/uid/TP40002644-BBCEACAJ
             #[cfg(feature = "usbportinfo-interface")]
-            interface: get_int_property(usb_device, "bInterfaceNumber", kCFNumberSInt8Type)
-                .map(|x| x as u8),
+            interface: get_int_property(usb_device, "bInterfaceNumber")
+                .map(|x| x as u8)
+                .ok(),
         })
     } else if get_parent_device_by_type(service, bluetooth_device_class_name).is_some() {
         SerialPortType::BluetoothPort
@@ -440,43 +398,15 @@ cfg_if! {
                         "IOServiceMatching returned a NULL dictionary.",
                     ));
                 }
-                let _classes_to_match_guard = scopeguard::guard((), |_| {
-                    CFRelease(classes_to_match as *const c_void);
-                });
+                let mut classes_to_match = CFMutableDictionary::wrap_under_create_rule(classes_to_match);
 
                 // Populate the search dictionary with a single key/value pair indicating that we're
                 // searching for serial devices matching the RS232 device type.
-                let key = CFStringCreateWithCString(
-                    kCFAllocatorDefault,
-                    kIOSerialBSDTypeKey,
-                    kCFStringEncodingUTF8,
-                );
-                if key.is_null() {
-                    return Err(Error::new(
-                        ErrorKind::Unknown,
-                        "Failed to allocate key string.",
-                    ));
-                }
-                let _key_guard = scopeguard::guard((), |_| {
-                    CFRelease(key as *const c_void);
-                });
-
-                let value = CFStringCreateWithCString(
-                    kCFAllocatorDefault,
-                    kIOSerialBSDAllTypes,
-                    kCFStringEncodingUTF8,
-                );
-                if value.is_null() {
-                    return Err(Error::new(
-                        ErrorKind::Unknown,
-                        "Failed to allocate value string.",
-                    ));
-                }
-                let _value_guard = scopeguard::guard((), |_| {
-                    CFRelease(value as *const c_void);
-                });
-
-                CFDictionarySetValue(classes_to_match, key as CFTypeRef, value as CFTypeRef);
+                let search_key = CStr::from_ptr(kIOSerialBSDTypeKey);
+                let search_key = CFString::from_static_string(search_key.to_str().map_err(|_| Error::new(ErrorKind::Unknown, "Failed to convert search key string"))?);
+                let search_value = CStr::from_ptr(kIOSerialBSDAllTypes);
+                let search_value = CFString::from_static_string(search_value.to_str().map_err(|_| Error::new(ErrorKind::Unknown, "Failed to convert search key string"))?);
+                classes_to_match.set(search_key, search_value);
 
                 // Get an interface to IOKit
                 let mut master_port: mach_port_t = MACH_PORT_NULL;
@@ -488,11 +418,17 @@ cfg_if! {
                     ));
                 }
 
-                // Run the search.
+                // Run the search. IOServiceGetMatchingServices consumes one reference count of
+                // classes_to_match, so explicitly retain.
+                //
+                // TODO: We could also just mem::forget classes_to_match like in
+                // TCFType::into_CFType. Is there a special reason that there is no
+                // TCFType::into_concrete_TypeRef()?
+                CFRetain(classes_to_match.as_CFTypeRef());
                 let mut matching_services = MaybeUninit::uninit();
                 kern_result = IOServiceGetMatchingServices(
                     kIOMasterPortDefault,
-                    CFRetain(classes_to_match as *const c_void) as *const __CFDictionary,
+                    classes_to_match.as_concrete_TypeRef(),
                     matching_services.as_mut_ptr(),
                 );
                 if kern_result != KERN_SUCCESS {
@@ -525,42 +461,29 @@ cfg_if! {
                         kCFAllocatorDefault,
                         0,
                     );
-                    let _props_guard = scopeguard::guard((), |_| {
-                        CFRelease(props.assume_init() as *const c_void);
-                    });
-
                     if result == KERN_SUCCESS {
+                        // A successful call to IORegistryEntryCreateCFProperties indicates that a
+                        // properties dict has been allocated and we as the caller are in charge of
+                        // releasing it.
+                        let props = props.assume_init();
+                        let props: CFDictionary<CFString, *const c_void> = CFDictionary::wrap_under_create_rule(props);
+
                         for key in ["IOCalloutDevice", "IODialinDevice"].iter() {
-                            let key_cstring = CString::new(*key).unwrap();
-                            let key_cfstring = CFStringCreateWithCString(
-                                kCFAllocatorDefault,
-                                key_cstring.as_ptr(),
-                                kCFStringEncodingUTF8,
-                            );
-                            let _key_cfstring_guard = scopeguard::guard((), |_| {
-                                CFRelease(key_cfstring as *const c_void);
-                            });
+                            let cf_key = CFString::new(key);
 
-                            let mut value = std::ptr::null();
-                            let found = CFDictionaryGetValueIfPresent(props.assume_init(), key_cfstring as *const c_void, &mut value);
-                            if found == true as Boolean {
-                                let type_id = CFGetTypeID(value);
-                                if type_id == CFStringGetTypeID() {
-                                    let mut buf = Vec::with_capacity(256);
-
-                                    CFStringGetCString(
-                                        value as CFStringRef,
-                                        buf.as_mut_ptr(),
-                                        buf.capacity() as isize,
-                                        kCFStringEncodingUTF8,
-                                    );
-                                    let path = CStr::from_ptr(buf.as_ptr()).to_string_lossy();
-                                    vec.push(SerialPortInfo {
-                                        port_name: path.to_string(),
-                                        port_type: port_type(modem_service),
-                                    });
-                                } else {
-                                    return Err(Error::new(ErrorKind::Unknown, "Found invalid type for TypeID"));
+                            if let Some(cf_ref) = props.find(cf_key) {
+                                let cf_type = CFType::wrap_under_get_rule(*cf_ref);
+                                match cf_type
+                                     .downcast::<CFString>()
+                                     .map(|s| s.to_string())
+                                {
+                                    Some(path) => {
+                                        vec.push(SerialPortInfo {
+                                            port_name: path,
+                                            port_type: port_type(modem_service),
+                                        });
+                                    }
+                                    None => return Err(Error::new(ErrorKind::Unknown, format!("Failed to get string value for {}", key))),
                                 }
                             } else {
                                 return Err(Error::new(ErrorKind::Unknown, format!("Key {} missing in dict", key)));
@@ -574,6 +497,14 @@ cfg_if! {
             Ok(vec)
         }
     } else if #[cfg(all(target_os = "linux", not(target_env = "musl"), feature = "libudev"))] {
+        fn is_rfcomm(device: &libudev::Device) -> bool {
+            device
+                .sysname()
+                .and_then(|o| o.to_str())
+                .map(|s| s.starts_with("rfcomm"))
+                .unwrap_or(false)
+        }
+
         /// Scans the system for serial ports and returns a list of them.
         /// The `SerialPortInfo` struct contains the name of the port
         /// which can be used for opening it.
@@ -584,22 +515,22 @@ cfg_if! {
                 enumerator.match_subsystem("tty")?;
                 let devices = enumerator.scan_devices()?;
                 for d in devices {
-                    if let Some(p) = d.parent() {
-                        if let Some(devnode) = d.devnode() {
-                            if let Some(path) = devnode.to_str() {
-                                if let Some(driver) = p.driver() {
-                                    if driver == "serial8250" && crate::new(path, 9600).open().is_err() {
-                                        continue;
-                                    }
+                    if let Some(devnode) = d.devnode().and_then(|o| o.to_str()) {
+                        let parent = d.parent();
+                        if parent.is_some() || is_rfcomm(&d) {
+                            if let Some(driver) = parent.as_ref().and_then(|d| d.driver()) {
+                                if driver == "serial8250" && crate::new(devnode, 9600).open().is_err() {
+                                    continue;
                                 }
-                                // Stop bubbling up port_type errors here so problematic ports are just
-                                // skipped instead of causing no ports to be returned.
-                                if let Ok(pt) = port_type(&d) {
-                                    vec.push(SerialPortInfo {
-                                        port_name: String::from(path),
-                                        port_type: pt,
-                                    });
-                                }
+                            }
+
+                            // Stop bubbling up port_type errors here so problematic ports are just
+                            // skipped instead of causing no ports to be returned.
+                            if let Ok(pt) = port_type(&d) {
+                                vec.push(SerialPortInfo {
+                                    port_name: String::from(devnode),
+                                    port_type: pt,
+                                });
                             }
                         }
                     }
@@ -612,29 +543,103 @@ cfg_if! {
         use std::io::Read;
         use std::path::Path;
 
+        fn is_rfcomm(path: &Path) -> bool {
+            path
+                .file_name()
+                .and_then(|o| o.to_str())
+                .map(|s| s.starts_with("rfcomm"))
+                .unwrap_or(false)
+        }
+
+        fn read_file_to_trimmed_string(dir: &Path, file: &str) -> Option<String> {
+            let path = dir.join(file);
+            let mut s = String::new();
+            File::open(path).ok()?.read_to_string(&mut s).ok()?;
+            Some(s.trim().to_owned())
+        }
+
+        fn read_file_to_u16(dir: &Path, file: &str) -> Option<u16> {
+            u16::from_str_radix(&read_file_to_trimmed_string(dir, file)?, 16).ok()
+        }
+
+        #[cfg(feature = "usbportinfo-interface")]
+        fn read_file_to_u8(dir: &Path, file: &str) -> Option<u8> {
+            u8::from_str_radix(&read_file_to_trimmed_string(dir, file)?, 16).ok()
+        }
+
+        fn read_port_type(path: &Path) -> Option<SerialPortType> {
+            let path = path
+                .canonicalize()
+                .ok()?;
+            let subsystem = path.join("subsystem").canonicalize().ok()?;
+            let subsystem = subsystem.file_name()?.to_string_lossy();
+
+            match subsystem.as_ref() {
+                // Broadcom SoC UARTs (of Raspberry Pi devices).
+                "amba" => Some(SerialPortType::Unknown),
+                "pci" => Some(SerialPortType::PciPort),
+                "pnp" => Some(SerialPortType::Unknown),
+                "usb" => usb_port_type(&path),
+                "usb-serial" => usb_port_type(path.parent()?),
+                _ => None,
+            }
+        }
+
+        fn usb_port_type(interface_path: &Path) -> Option<SerialPortType> {
+            let info = read_usb_port_info(interface_path)?;
+            Some(SerialPortType::UsbPort(info))
+        }
+
+        fn read_usb_port_info(interface_path: &Path) -> Option<UsbPortInfo> {
+            let device_path = interface_path.parent()?;
+
+            let vid = read_file_to_u16(&device_path, "idVendor")?;
+            let pid = read_file_to_u16(&device_path, "idProduct")?;
+            #[cfg(feature = "usbportinfo-interface")]
+            let interface = read_file_to_u8(&interface_path, &"bInterfaceNumber");
+            let serial_number = read_file_to_trimmed_string(&device_path, &"serial");
+            let product = read_file_to_trimmed_string(&device_path, &"product");
+            let manufacturer = read_file_to_trimmed_string(&device_path, &"manufacturer");
+
+            Some(UsbPortInfo {
+                vid,
+                pid,
+                serial_number,
+                manufacturer,
+                product,
+                #[cfg(feature = "usbportinfo-interface")]
+                interface,
+            })
+        }
+
         /// Scans `/sys/class/tty` for serial devices (on Linux systems without libudev).
         pub fn available_ports() -> Result<Vec<SerialPortInfo>> {
             let mut vec = Vec::new();
             let sys_path = Path::new("/sys/class/tty/");
-            let device_path = Path::new("/dev");
-            let mut s;
+            let dev_path = Path::new("/dev");
             for path in sys_path.read_dir().expect("/sys/class/tty/ doesn't exist on this system") {
                 let raw_path = path?.path().clone();
                 let mut path = raw_path.clone();
 
-                path.push("device");
-                if !path.is_dir() {
-                    continue;
-                }
-
-                path.push("driver_override");
-                if path.is_file() {
-                    s = String::new();
-                    File::open(path)?.read_to_string(&mut s)?;
-                    if &s == "(null)\n" {
+                let port_type = if is_rfcomm(&raw_path) {
+                    SerialPortType::BluetoothPort
+                } else {
+                    path.push("device");
+                    if !path.is_dir() {
                         continue;
                     }
-                }
+
+                    // Determine port type and proceed, if it's a known.
+                    //
+                    // TODO: Switch to a likely more readable let-else statement when our MSRV supports
+                    // it.
+                    let port_type = read_port_type(&path);
+                    if let Some(port_type) = port_type {
+                        port_type
+                    } else {
+                        continue;
+                    }
+                };
 
                 // Generate the device file path `/dev/DEVICE` from the TTY class path
                 // `/sys/class/tty/DEVICE` and emit a serial device if this path exists. There are
@@ -642,14 +647,14 @@ cfg_if! {
                 //
                 // See https://github.com/serialport/serialport-rs/issues/66 for details.
                 if let Some(file_name) = raw_path.file_name() {
-                    let device_file = device_path.join(file_name);
+                    let device_file = dev_path.join(file_name);
                     if !device_file.exists() {
                         continue;
                     }
 
                     vec.push(SerialPortInfo {
                         port_name: device_file.to_string_lossy().to_string(),
-                        port_type: SerialPortType::Unknown,
+                        port_type,
                     });
                 }
             }