+++ src/windows/com.rs
@@ -88,6 +88,11 @@ impl COMPort {
         dcb::set_flow_control(&mut dcb, builder.flow_control);
         dcb::set_dcb(handle, dcb)?;
 
+        // Try to set DTR on best-effort.
+        if let Some(dtr) = builder.dtr_on_open {
+            let _ = com.write_data_terminal_ready(dtr);
+        }
+
         com.set_timeout(builder.timeout)?;
         com.port_name = Some(builder.path.clone());
         Ok(com)